+++ picocli-annotation-processing-tests/src/test/java/picocli/annotation/processing/tests/Issue777Test.java
@@ -12,7 +12,7 @@ import static com.google.testing.compile.CompilationSubject.assertThat;
 import static com.google.testing.compile.Compiler.javac;
 
 public class Issue777Test {
-    @Ignore
+//    @Ignore
     @Test
     public void testIssue777() {
         CommandSpec2YamlProcessor processor = new CommandSpec2YamlProcessor();
@@ -23,24 +23,39 @@ public class Issue777Test {
                                 "picocli/issue777/MutuallyExclusiveOptionsDemo.java"));
 
         assertThat(compilation).succeeded();
+
+        processor.strings.forEach(System.out::println);
+    }
+
+//    @Ignore
+    @Test
+    public void testIssue777Composite() {
+        CommandSpec2YamlProcessor processor = new CommandSpec2YamlProcessor();
+        Compilation compilation =
+                javac()
+                        .withProcessors(processor)
+                        .compile(JavaFileObjects.forResource(
+                                "picocli/issue777/CompositeGroupDemo.java"));
+
+        assertThat(compilation).succeeded();
     }
 
 //    @Test
-//    public void testIssue769Details() {
+//    public void testIssue777Details() {
 //        Processor processor = new AnnotatedCommandSourceGeneratorProcessor();
 //        Compilation compilation =
 //                javac()
 //                        .withProcessors(processor)
 //                        .compile(JavaFileObjects.forResource(
-//                                "picocli/issue769/MyMixin.java"));
+//                                "picocli/issue777/CompositeGroupDemo.java"));
 //
 //        assertThat(compilation).succeeded();
 //        assertThat(compilation)
-//                .generatedFile(StandardLocation.SOURCE_OUTPUT, "generated/picocli/issue769/MyMixin.java")
-//                .hasSourceEquivalentTo(JavaFileObjects.forResource("generated/picocli/issue769/MyMixin.java"));
+//                .generatedFile(StandardLocation.SOURCE_OUTPUT, "generated/picocli/issue777/CompositeGroupDemo.java")
+//                .hasSourceEquivalentTo(JavaFileObjects.forResource("generated/picocli/issue777/CompositeGroupDemo.java"));
 //
 //        assertThat(compilation)
-//                .generatedFile(StandardLocation.SOURCE_OUTPUT, "generated/picocli/issue769/SubCommand.java")
-//                .hasSourceEquivalentTo(JavaFileObjects.forResource("generated/picocli/issue769/SubCommand.java"));
+//                .generatedFile(StandardLocation.SOURCE_OUTPUT, "generated/picocli/issue777/SubCommand.java")
+//                .hasSourceEquivalentTo(JavaFileObjects.forResource("generated/picocli/issue777/SubCommand.java"));
 //    }
 }